+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Added
 ~~~~~
 - ``run_tests.py`` script to execute all unittests
 - PDB-redo download for structures
+Changed
+~~~~~~~
+- Removed reference to deprecated module ``iotbx.pdb.mining``
 
 0.1.0
 -----